add IsDeleted property and Delete method for Persons update finacial field/property to be decimal for Persons

+++ PortoFlipPizzaSystem/ProtoFlipPizzaSystem.Models/Administrator/Contracts/ICustomer.cs
@@ -8,7 +8,7 @@ namespace ProtoFlipPizzaSystem.Models.Administrator.Contracts
 {
     public interface ICustomer : IPerson
     {
-        double Discount { get; }
+        decimal Discount { get; }
         int NumberOfOrders { get; }
     }
 }
